rewrote chwdb-cli.sh w/ my script template to enable specifying many options
this will allow it to be used to login as the mariadb root user, or to invoke bash in the container instead of the db cli.

chwdb-cli.sh --cmd=bash -q
chwdb-cli.sh --user=root --pswd=rtpswd --db=""

It's also possible to pass in some sql commands as long as you remember to exit.
echo "SHOW TABLES; exit;" | chwdb-cli.sh --quiet