Added wine lookup tables, replace case unit conversion factor column
New wine lookup tables: countries, regions, subregions, appellations.

replace those name fields in the Wines table w/ the lookup Ids.

Reordered LookupCaseUnits columns and replaced the `LabelVolumeConvFactor` column with `VolumeInLabelUnits` which is the value that is usually needed, and the conversion factor to/from ml can be calculated from that and `VolumeInMilliliters` if needed.